Make blockProc work for GPU processing
It now first transfers a subset of a large volume to a staging
block on the host. Then, the host block is tranfered to the GPU,
and the supplied function is called with the GPU blocks. Finally,
the same happens in reverse for output.

This also necessitated updating the tests.